Fixed exception when destructing

+++ src/object.cpp
@@ -5,9 +5,11 @@ Object::Object(glm::vec3 pos, glm::vec3 rot, glm::vec3 scl){
 }
 
 Object::~Object(){
-	std::vector<Object*>* vec = &parent->components; // Get a simplified pointer to the parents components
-	vec->erase(std::remove(vec->begin(), vec->end(), this), vec->end()); // Hunt down and kill itself from the parent vector
-	parent = NULL;
+	if (parent != NULL){
+		std::vector<Object*>* vec = &parent->components; // Get a simplified pointer to the parents components
+		vec->erase(std::remove(vec->begin(), vec->end(), this), vec->end()); // Hunt down and kill itself from the parent vector
+		parent = NULL;
+	}
 }
 
 void Object::adopt(Object* C){
